# api/openapi.yaml
openapi: 3.0.3
info:
  title: Travel Companion API
  version: 1.0.0
  description: |
    Travel Companion — connect travelers who need language assistance with
    those who can provide it on the same flight.
  contact:
    name: Travel Companion
servers:
  - url: http://localhost:5000
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [first_name, languages_spoken]
      properties:
        first_name:
          type: string
          maxLength: 100
        languages_spoken:
          type: array
          items:
            type: string
          description: ISO 639-1 codes
    UserProfile:
      type: object
      properties:
        id:
          type: string
        first_name:
          type: string
        languages_spoken:
          type: array
          items:
            type: string
        mobile_verified:
          type: boolean
        is_blocked:
          type: boolean
        notification_opt_out:
          type: boolean
        created_at:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        first_name:
          type: string
          maxLength: 100
        languages_spoken:
          type: array
          items:
            type: string
        notification_opt_out:
          type: boolean
...
    SendMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
          maxLength: 2000
    MessageResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        thread_id:
          type: string
          format: uuid
        sender_id:
          type: string
        body:
          type: string
        is_read:
          type: boolean
        sent_at:
          type: string
          format: date-time
...
    ReportRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          maxLength: 500
...
    AirportDto:
      type: object
      properties:
        iata_code:
          type: string
        name:
          type: string
        city:
          type: string
        country:
          type: string
        timezone:
          type: string
    FlightDto:
      type: object
      properties:
        flight_number:
          type: string
        airline:
          type: string
        origin_iata:
          type: string
        destination_iata:
          type: string
        departure_time:
          type: string
          format: date-time
        arrival_time:
          type: string
          format: date-time
...
    ReportDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reporter_id:
          type: string
        reason:
          type: string
        created_at:
          type: string
          format: date-time
        dismissed_at:
          type: string
          format: date-time
          nullable: true
    ClientErrorRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
        stack:
          type: string
        url:
          type: string
        userAgent:
          type: string
...
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Create user profile after Auth0 signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '409':
          description: Profile already exists
        '422':
          description: Validation error
  /auth/sync:
    post:
      tags: [Auth]
      summary: Sync mobile_verified status from Auth0 JWT
      responses:
        '200':
          description: Synced
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: User profile not found
  /users/me:
    get:
      tags: [Users]
      summary: Get own profile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: Not found
    patch:
      tags: [Users]
      summary: Update own profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
    delete:
      tags: [Users]
      summary: Delete own account
      responses:
        '204':
          description: Deleted
  /users/{id}/report:
    post:
      tags: [Reports]
      summary: Report a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '201':
          description: Report created
        '409':
          description: Already reported
  /users/{id}/block:
    post:
      tags: [Blocks]
      summary: Block a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Blocked
        '409':
          description: Already blocked
...
  /users/blocked/{id}:
    delete:
      tags: [Blocks]
      summary: Unblock a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Unblocked
...
  /messages/threads/{id}:
    get:
      tags: [Messages]
      summary: Get messages in a thread
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageResponse'
    post:
      tags: [Messages]
      summary: Send a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '201':
          description: Sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /flights/search:
    get:
      tags: [Flights]
      summary: Search flights (proxy to AviationStack)
      parameters:
        - name: origin_iata
          in: query
          required: true
          schema:
            type: string
        - name: destination_iata
          in: query
          required: true
          schema:
            type: string
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FlightDto'
  /airports/search:
    get:
      tags: [Airports]
      summary: IATA code autocomplete (unauthenticated)
      security: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AirportDto'
  /logs/client-error:
    post:
      tags: [Logs]
      summary: Log a client-side error
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientErrorRequest'
      responses:
        '204':
          description: Logged
        '413':
          description: Payload too large
...
  /admin/users/{id}/reports:
    get:
      tags: [Admin]
      summary: Get reports against a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportDto'
  /admin/users/{id}/block:
    post:
      tags: [Admin]
      summary: Block a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Blocked
  /admin/users/{id}/unblock:
    post:
      tags: [Admin]
      summary: Unblock a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Unblocked
  /admin/users/{id}/reports/dismiss:
    post:
      tags: [Admin]
      summary: Dismiss all reports for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Dismissed
...